Refactor el controlador de descargas para importar procesar_albumes desde album_postprocessor y añade la opción '--verbose' en el comando yt-dlp. Cambia el nombre del contenedor en docker-compose.yml a yt_subs.
